Fix the signature of ves_icall_System_Runtime_InteropServices_Marshal_AllocCoTaskMemSize.

+++ mono/metadata/marshal.h
@@ -738,7 +738,7 @@ ves_icall_System_Runtime_InteropServices_Marshal_AllocCoTaskMem (int size, MonoE
 
 ICALL_EXPORT
 void*
-ves_icall_System_Runtime_InteropServices_Marshal_AllocCoTaskMemSize (gulong size, MonoError *error);
+ves_icall_System_Runtime_InteropServices_Marshal_AllocCoTaskMemSize (gsize size, MonoError *error);
 
 ICALL_EXPORT
 void
